Adds fix option to xfmt
... to make it easy to fix formatting that is tedious to fix by hand.

Changes terminology from "checker" to "formatter" to better reflect that both
checking and fixing is supported.

Rewrites JSON formatter to make it more useful outside the context of a
formatter plugin.

Changes json feedback to optionally include the path of the file being
processed to make it easier to read when multiple files are being processed.

Changes how feedback is printed to fix problem whereby two chunks would overlap
(hence a newline every time) and to fix problem whereby empty feedback would
also be printed (apparent now that each feedback is properly terminated with a
newline).

Adds test checking the plugins detected to make it immediately obvious when
this is the cause of tests failing.

S