Adds python formatting
... to reduce the number tools that must be explicitly managed to control the
    code quality in a poroject.

Integrates with black using `format_file_contents` to maintain control of
printing to stdout. A side effect is that it allows the same feedback format
to be used for python and json. Other than that the plugin largely replicates
`format_file_in_place`.

Moves samples that are expected to be bad from source that should be inspected
because `__pychace__` kept appearing and I cannot be which messes up the
collection test and fixing that will have to wait.

Adds python samples to test collection and formatting.